fix: media gallery fixtures, missing news list and detail in page content
