+++ demo/api/src/documents/pages/blocks/page-content.block.ts
@@ -2,7 +2,7 @@ import { BaseBlocksBlockItemData, BaseBlocksBlockItemInput, BlockField, createBl
 import { AnchorBlock, DamImageBlock } from "@comet/cms-api";
 import { AccordionBlock } from "@src/common/blocks/accordion.block";
 import { LinkListBlock } from "@src/common/blocks/link-list.block";
-import { MediaGalleryListBlock } from "@src/common/blocks/media-gallery.block";
+import { MediaGalleryBlock } from "@src/common/blocks/media-gallery.block";
 import { RichTextBlock } from "@src/common/blocks/rich-text.block";
 import { SpaceBlock } from "@src/common/blocks/space.block";
 import { StandaloneCallToActionListBlock } from "@src/common/blocks/standalone-call-to-action-list.block";
@@ -14,6 +14,8 @@ import { ColumnsBlock } from "@src/documents/pages/blocks/columns.block";
 import { ContentGroupBlock } from "@src/documents/pages/blocks/content-group.block";
 import { KeyFactsBlock } from "@src/documents/pages/blocks/key-facts.block";
 import { TeaserBlock } from "@src/documents/pages/blocks/teaser.block";
+import { NewsDetailBlock } from "@src/news/blocks/news-detail.block";
+import { NewsListBlock } from "@src/news/blocks/news-list.block";
 import { UserGroup } from "@src/user-groups/user-group";
 import { IsEnum } from "class-validator";
 
@@ -32,7 +34,12 @@ export const mediaBlocks = {
     image: DamImageBlock,
     fullWidthImage: FullWidthImageBlock,
     media: StandaloneMediaBlock,
-    mediaGallery: MediaGalleryListBlock,
+    mediaGallery: MediaGalleryBlock,
+};
+
+export const newsBlocks = {
+    newsDetail: NewsDetailBlock,
+    newsList: NewsListBlock,
 };
 
 export const navigationBlocks = {
